don't use size_t instead of unsigned long (it's not the same on some 32 bits systems like Fedora 4)
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/aims/aims-free/branches/4.4@59232 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ aimsdata/src/aimsdata/data/data.cc
@@ -57,7 +57,7 @@ INSTANTIATE_GENERIC_OBJECT_TYPE( _TMP_ )
 #undef _TMP_
 #ifndef __LP64__
 INSTANTIATE_GENERIC_OBJECT_TYPE( AimsData<long> )
-INSTANTIATE_GENERIC_OBJECT_TYPE( AimsData<size_t> )
+INSTANTIATE_GENERIC_OBJECT_TYPE( AimsData<unsigned long> )
 #endif
 
 INSTANTIATE_GENERIC_OBJECT_TYPE( rc_ptr< AimsData<int8_t> > )
@@ -81,7 +81,7 @@ INSTANTIATE_GENERIC_OBJECT_TYPE( _TMP_ )
 #undef _TMP_
 #ifndef __LP64__
 INSTANTIATE_GENERIC_OBJECT_TYPE( rc_ptr< AimsData<long> > )
-INSTANTIATE_GENERIC_OBJECT_TYPE( rc_ptr< AimsData<size_t> > )
+INSTANTIATE_GENERIC_OBJECT_TYPE( rc_ptr< AimsData<unsigned long> > )
 #endif
 
 } // namespace carto
